Fix AI response route and PDF extraction
