+++ app/utils/extractPdfText.ts
@@ -38,10 +38,7 @@ export async function extractPdfText(buffer: Buffer): Promise<string> {
  * Fallback OCR if pdfjs can't extract text (e.g. scanned document)
  */
 async function extractWithOCR(buffer: Buffer): Promise<string> {
-  const worker = await createWorker();
-  await worker.load();
-  await worker.loadLanguage('eng');
-  await worker.initialize('eng');
+  const worker = await createWorker('eng');
 
   let ocrText = '';
 
@@ -54,7 +51,7 @@ async function extractWithOCR(buffer: Buffer): Promise<string> {
       const viewport = page.getViewport({ scale: 2 });
 
       const canvas = createCanvas(viewport.width, viewport.height);
-      const ctx = canvas.getContext('2d');
+      const ctx = canvas.getContext('2d') as any;
       await page.render({ canvasContext: ctx, viewport }).promise;
 
       const imageBuffer = canvas.toBuffer('image/png');
