+++ package.json
@@ -21,6 +21,7 @@
     "mammoth": "1.9.1",
     "next": "15.3.3",
     "next-auth": "4.24.11",
+    "openai": "^5.7.0",
     "pdf-lib": "1.17.1",
     "pdfjs-dist": "^3.11.174",
     "prisma": "5.14.0",
@@ -30,6 +31,7 @@
   },
   "devDependencies": {
     "@tailwindcss/postcss": "4.0.0",
+    "@types/bcrypt": "^5.0.2",
     "@types/node": "20.0.0",
     "@types/react": "19.0.0",
     "@types/react-dom": "19.0.0",
